refactored around `preproc`, `symmetry_function` and so on
defined instance variables of `PCA` correctly
changed `load_xyz` and `load_poscars` as non-class method,
    and moved to `file_io.py`
  changed `preproc` to outer scope of both
  both returns datasets and elements
renamed `SymmetryFunctionDatasetGenerator` to `DatasetGenerator`,
    and moved to `utils.py` along with `scatter_dataset`
  it is no longer dependent on `SymmetryFunctionDataset`
  requires datasets in initialization
aligned codes of main() function in `hdnnp.py`
