+++ hdnnpy/util.py
@@ -111,14 +111,14 @@ def dump_lammps(file_path, preproc, masters):
 
 
 def dump_training_result(file_path, result):
-    args = {k:v if not isinstance(v, Path) else str(v)
-            for k,v in vars(stg.args).items() if not k.startswith('_')}
-    file = {k:v if not isinstance(v, Path) else str(v)
-            for k,v in vars(stg.file).items() if not k.startswith('_')}
-    dataset = {k:v if not isinstance(v, Path) else str(v)
-               for k,v in vars(stg.dataset).items() if not k.startswith('_')}
-    model = {k:v if not isinstance(v, Path) else str(v)
-             for k,v in vars(stg.model).items() if not k.startswith('_')}
+    args = {k: v if not isinstance(v, Path) else str(v)
+            for k, v in vars(stg.args).items() if not k.startswith('_')}
+    file = {k: v if not isinstance(v, Path) else str(v)
+            for k, v in vars(stg.file).items() if not k.startswith('_')}
+    dataset = {k: v if not isinstance(v, Path) else str(v)
+               for k, v in vars(stg.dataset).items() if not k.startswith('_')}
+    model = {k: v if not isinstance(v, Path) else str(v)
+             for k, v in vars(stg.model).items() if not k.startswith('_')}
 
     with file_path.open('w') as f:
         yaml.dump({
